LPS-60781 separates RowCheckers for users
