+++ portal-impl/src/com/liferay/portlet/rolesadmin/search/UnsetUserRoleChecker.java
@@ -20,34 +20,19 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Role;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.security.membershippolicy.RoleMembershipPolicyUtil;
-import com.liferay.portal.kernel.service.UserLocalServiceUtil;
 
 import javax.portlet.RenderResponse;
 
 /**
  * @author Brian Wing Shun Chan
+ * @author Drew Brokke
  */
-public class UserRoleChecker extends EmptyOnClickRowChecker {
+public class UnsetUserRoleChecker extends EmptyOnClickRowChecker {
 
-	public UserRoleChecker(RenderResponse renderResponse, Role role) {
+	public UnsetUserRoleChecker(RenderResponse renderResponse, Role role) {
 		super(renderResponse);
 
 		_role = role;
-	}
-
-	@Override
-	public boolean isChecked(Object obj) {
-		User user = (User)obj;
-
-		try {
-			return UserLocalServiceUtil.hasRoleUser(
-				_role.getRoleId(), user.getUserId());
-		}
-		catch (Exception e) {
-			_log.error(e, e);
-
-			return false;
-		}
 	}
 
 	@Override
@@ -55,10 +40,7 @@ public class UserRoleChecker extends EmptyOnClickRowChecker {
 		User user = (User)obj;
 
 		try {
-			if (isChecked(user) ||
-				RoleMembershipPolicyUtil.isRoleRequired(
-					user.getUserId(), _role.getRoleId()) ||
-				!RoleMembershipPolicyUtil.isRoleAllowed(
+			if (RoleMembershipPolicyUtil.isRoleRequired(
 					user.getUserId(), _role.getRoleId())) {
 
 				return true;
@@ -72,7 +54,7 @@ public class UserRoleChecker extends EmptyOnClickRowChecker {
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		UserRoleChecker.class);
+		UnsetUserRoleChecker.class);
 
 	private final Role _role;
